nodes/mc/gomc/libmango/py/requirements.txt
ply==3.10
lxml==3.8.0
PyYAML==3.12
pyzmq==17.0.0b1
